split zarr.hpp into zarr.hpp and zarr.cpp
zarr.hpp took forever to compile due to lots of inline code so split it in two
